Remove beta button from grid, add feedback button to data section linking to collaboration maze
